RAC-16: Move sources to src
