+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/workspaces/shared/package.json
@@ -2,5 +2,8 @@
   "name": "@akeneo-pim-community/shared",
   "version": "1.0.0",
   "private": true,
-  "main": "index.ts"
+  "main": "index.ts",
+  "paths": {
+    "@src/*": ["src/*"]
+  }
 }
