+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/workspaces/shared/src/components/index.ts
