+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/workspaces/shared/src/hooks/useAkeneoTheme.ts
