+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/workspaces/shared/src/icons/index.ts
