+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/workspaces/shared/src/theme/AkeneoThemeProvider.tsx
