+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/workspaces/shared/src/theme/theme.ts
